+++ BloodBank/src/main/java/com/example/BloodBank/service/service_interface/ICRUDService.java
@@ -3,7 +3,7 @@ package com.example.BloodBank.service.service_interface;
 public interface ICRUDService<T> {
 
     void Create(T entity) throws Exception;
-    T Read(int id) throws Exception;
+    T Read(long id) throws Exception;
     void Update(T entity) throws Exception;
     void Delete(T entity) throws Exception;
